fix(windows): cast DWMWA_* attribute consts to u32 for DwmSetWindowAttribute
windows-sys types the DWMWA_CAPTION_COLOR / DWMWA_TEXT_COLOR consts as i32 but
DwmSetWindowAttribute's attribute parameter as u32, so v0.2.3 failed to compile
on the Windows runners (E0308). Cast the attribute to u32. Windows-only block;
the rest of the call already type-checked per the failed-build log.

+++ src-tauri/src/main.rs
@@ -485,7 +485,12 @@ fn set_titlebar(window: tauri::WebviewWindow, bg: String, text: String, dark: bo
         };
         if let Ok(hwnd) = window.hwnd() {
             let handle: HWND = hwnd.0 as _;
-            for (attr, hex) in [(DWMWA_CAPTION_COLOR, &bg), (DWMWA_TEXT_COLOR, &text)] {
+            // windows-sys types the DWMWA_* consts as i32 but the attribute
+            // parameter as u32, so cast.
+            for (attr, hex) in [
+                (DWMWA_CAPTION_COLOR as u32, &bg),
+                (DWMWA_TEXT_COLOR as u32, &text),
+            ] {
                 if let Some(c) = hex_colorref(hex) {
                     unsafe {
                         let _ = DwmSetWindowAttribute(
